refactor: don't use index import, circular dependency
